typeck: move `check_pat_walk` and children to `pat.rs`.
